+++ src/test/resources/org/sonar/plugins/delphi/symbol/helpers/ClassHelperOverride.pas
@@ -1,0 +1,33 @@
+unit ClassHelperOverride;
+
+{This is a sample Delphi file.}
+
+interface
+
+type
+  TFoo = class
+    procedure Bar;
+  end;
+
+  TFooHelper = class helper for TFoo
+    procedure Bar;
+  end;
+
+implementation
+
+procedure TFoo.Bar;
+begin
+  // Do nothing
+end;
+
+procedure TFooHelper.Bar;
+begin
+  // Do nothing
+end;
+
+procedure Test(Foo: TFoo);
+begin
+  Foo.Bar;
+end;
+
+end.